Print the First 98 Fibonacci Numbers

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,11 +9,11 @@
 int main(void)
 {
 	int count = 0;
-	double res = 1, no1 = 1, no2 = 1;
+	long double res = 1, no1 = 1, no2 = 1;
 
 	while (count < 98)
 	{
-		printf("%.0f", res);
+		printf("%.0Lf", res);
 		res = no1 + no2;
 		no2 = no1;
 		no1 = res;
